requirements.txt
openai>=1.30.0
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.129.0
tweepy>=4.14.0
Pillow>=10.0.0
schedule>=1.2.0
